% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{age_strat_fit_mod}
\alias{age_strat_fit_mod}
\title{Model to feed to `rootSolve::multiroot()` to estimate key parameters of the age stratified model based on observed inputs}
\usage{
age_strat_fit_mod(x, I, W_A, prev_A, H_A, W_C, prev_C, H_C, pars)
}
\arguments{
\item{x}{input estimates of parameters beta, N_eq, alpha, and omega (IN THAT ORDER)}

\item{I}{infected snail prevalence}

\item{W_A}{estimated mean worm burden in adult population}

\item{prev_A}{estimated prevalence in adult population}

\item{H_A}{adult population size}

\item{W_C}{estimated mean worm burden in child population}

\item{prev_C}{estimated prevalence in child population}

\item{H_C}{child population size}

\item{parameters}{parameter set of other key parameters}
}
\value{
Vector of four equation solutions
}
\description{
We have four unknown parameters and four solvable equations. This function
takes inputs and puts them into equations to estimate the value of these four unkown parameters,
alpha, beta, omega, and N_eq, from inputs: adult infection (prevalence and intensity),
child infection (prevalence and intensity), child and adult population sizes, infected snail prevalence
}
