% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{age_strat_get_pars}
\alias{age_strat_get_pars}
\title{Function taking `age_strat_fit_mod` and implementing within `rootSolve::multiroot()`
to return estimates of key parameters of the age stratified model based on observed inputs}
\usage{
age_strat_get_pars(f = age_strat_fit_mod, start = c(0.01, 1000, 1e-06,
  4), I, W_A, prev_A, H_A, W_C, prev_C, H_C, pars, tolerance = 1e-06)
}
\arguments{
\item{f}{function with equations to feed to `multiroot` function. Defaults to `age_strat_fit_mod`}

\item{start}{input estimates of parameters beta, N_eq, alpha, and omega (IN THAT ORDER)}

\item{I}{infected snail prevalence}

\item{W_A}{estimated mean worm burden in adult population}

\item{prev_A}{estimated prevalence in adult population}

\item{H_A}{adult population size}

\item{W_C}{estimated mean worm burden in child population}

\item{prev_C}{estimated prevalence in child population}

\item{H_C}{child population size}

\item{pars}{parameter set of other key parameters}

\item{tolerance}{minimum error to tolerate, else return error}
}
\value{
Estimate of porportion of snail population in E compartment
}
\description{
We have four unknown parameters and four solvable equations. This function
takes inputs and puts them into equations to estimate the value of these four unkown parameters,
alpha, beta, omega, and N_eq, from inputs: adult infection (prevalence and intensity),
child infection (prevalence and intensity), child and adult population sizes, infected snail prevalence
}
