% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{augment_age_strat_parameters}
\alias{augment_age_strat_parameters}
\title{Augment parameter set for age stratified schisto model with fit transmission and observed demographic/infection parameters}
\usage{
augment_age_strat_parameters(I, W_A, prev_A, H_A, cvrg_A, W_C, prev_C, H_C,
  cvrg_C, pars)
}
\arguments{
\item{I}{infected snail prevalence}

\item{W_A}{estimated mean worm burden in adult population}

\item{prev_A}{estimated prevalence in adult population}

\item{H_A}{adult population size}

\item{cvrg_A}{MDA coverage (proportion) among the adult population}

\item{W_C}{estimated mean worm burden in child population}

\item{prev_C}{estimated prevalence in child population}

\item{H_C}{child population size}

\item{cvrg_C}{MDA coverage (proportion) among the child population}

\item{pars}{parameter set of other key parameters}
}
\value{
parameter set with all parameters needed to run the age stratified model
}
\description{
Takes as input a few observed variables, uses `age_strat_get_pars` to estimate transmission parameters, adds them to the
base parameters, and returns the resulting full parameter set
}
