% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_dynamic_models.R
\name{base_mod}
\alias{base_mod}
\title{Basic schistosomiasis model}
\usage{
base_mod(t, n, parameters)
}
\arguments{
\item{t}{Vector of timepoints to return state variable estiamtes}

\item{n}{Vector of state variable initial conditions}

\item{parameters}{Named vector of model parameters}
}
\value{
A matrix of the state variables at all requested time points
}
\description{
A dynamic schistosomiasis model with negative (crowding induced reductions in fecundity)
and positive (mating limitation) density dependencies and functionality
to simulate mass drug administration, snail control, and other interventions.
}
