% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{beta_from_eggs}
\alias{beta_from_eggs}
\title{Function to estimate beta from egg output and other parameters}
\usage{
beta_from_eggs(egg_output, H, Lambda, N_eq, U, v, omega)
}
\arguments{
\item{egg_output}{mean eggs produced per 10mL per person (observed diagnostic data)}

\item{H}{number of individuals in the human population}

\item{Lambda}{snail force of infection}

\item{N_eq}{snail population size}

\item{U}{mean urine produced per day per individual in units of 10mL}

\item{v}{mean egg viability (probability an egg produced successfully hatches into miracidia)}

\item{omega}{human exposure/contamination parameter that regulates proportion of eggs produced that interact with snail population}
}
\value{
estimate of the per miracidial contact probability of infection, beta
}
\description{
Function to estimate beta from egg output and other parameters
}
