% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{cheever_eggs_to_females}
\alias{cheever_eggs_to_females}
\title{Cheever eggs to females function}
\usage{
cheever_eggs_to_females(eggs)
}
\arguments{
\item{eggs}{observed egg burden in eggs/10mL urine}
}
\value{
estimate of fecund female worms from egg burden
}
\description{
Uses reported log-log relationship between eggs in urine and female worms to estimate worm burden from egg burden
}
