% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{cvrg_from_pars}
\alias{cvrg_from_pars}
\title{Estimate MDA coverage necessary to reach breakpoint as a function of parameters
used to esimate breakpoint, density dependencies,}
\usage{
cvrg_from_pars(W_t, W_u, epsilon, pars, PDD = phi_Wk, DDF = rho_Wk,
  DDI = nil_1)
}
\arguments{
\item{W_t}{mean worm burden prior to MDA in treated population}

\item{W_u}{mean worm burden in untreated population}

\item{epsilon}{MDA efficacy (proportion of worms cleared from those treated)}

\item{pars}{additional parameters}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}
}
\value{
estimate of coverage required to reach breakpoint
}
\description{
Estimate MDA coverage necessary to reach breakpoint as a function of parameters
used to esimate breakpoint, density dependencies,
}
