% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{cvrg_from_W_bp}
\alias{cvrg_from_W_bp}
\title{Estimate MDA coverage necessary to reach breakpoint as a function of infected snail prevalence and parameters}
\usage{
cvrg_from_W_bp(W_t, W_bp, epsilon)
}
\arguments{
\item{W_t}{mean worm burden prior to MDA}

\item{W_bp}{breakpoint mean worm burden}

\item{epsilon}{MDA efficacy (proportion of worms cleared from those treated)}
}
\value{
estimate of MDA coverage necessary to reach breakpoint
}
\description{
Estimate MDA coverage necessary to reach breakpoint as a function of infected snail prevalence and parameters
}
