% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{egg_to_worm_fx}
\alias{egg_to_worm_fx}
\title{Function used in multiroot to estimate worm burden and clumping parameter from egg burden, prevalence and parameters}
\usage{
egg_to_worm_fx(x, parms)
}
\arguments{
\item{x}{initial estimates for the mean worm burden and clumping parameter, respectively}

\item{parms}{input egg burden (measured from surveys) measured in eggs/10mL for the population group}
}
\value{
Value of the two equations given inputs, but is mostly irrelevant other than its use within multiroot (see function `convert_burden_egg_to_worm`)
}
\description{
Function which uses equations representing 1) an estimate of prevalence given worm burden and clumping parameter and
2) estimated mean egg output per 10mL urine given worm burden, mating probability, DD fecundity, and peak egg release per 10mL
to estimate the worm burden and clumping parameter from input egg burden and prevalence
}
