% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{eggs_kap_get_W}
\alias{eggs_kap_get_W}
\title{Function to estimate worm burden given mean egg burden and dispersion of egg burden}
\usage{
eggs_kap_get_W(eggs, kap, zeta, m)
}
\arguments{
\item{eggs}{mean egg output}

\item{kap}{disperion parameter on egg output}

\item{zeta}{fecundity reduction parameter}

\item{m}{mean egg output per mated female worm with no fecundity reduction}
}
\value{
estimate of the mean worm burden
}
\description{
Uses `uniroot` to solve for W assuming mean egg burden = 0.5Wphi(W,k)rho(W,k)m
}
