% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{eq_W_get_alpha}
\alias{eq_W_get_alpha}
\title{Function to estimate probability of worm establishment per cercarial exposure from equilibrium worm burden}
\usage{
eq_W_get_alpha(W_C, omega_c, mu_W, mu_H_C, theta, I_P, N_eq)
}
\arguments{
\item{W_C}{equilibrium mean worm burden in the child population}

\item{omega_c}{contamination/exposure coefficient for children}

\item{mu_W}{daily mortality rate of adult worms}

\item{mu_H_C}{mortality rate of children}

\item{theta}{daily cercarial shedding rate of infected snails}

\item{I_P}{infected snail prevalence}

\item{N_eq}{equilibrium snail population size}
}
\description{
Assuming observed worm burden prior to any intervention is at euilibirum, we can estimate the worm acquisition rate
as a function of this equilibirum worm burden and host and worm turnover (mortality) rates. The only component
of the worm acquisition rate that is unkown is the probability of worm establishment per cercarial exposure, alpha, which we can estimate
from equilibirum mean worm burden and known parameters
}
