% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{eq_Ws_get_Omega_Lambda0}
\alias{eq_Ws_get_Omega_Lambda0}
\title{Function to estimate miracidial invasion rate and relative contamination coefficient from equilibirum egg burden and prevalence inputs}
\usage{
eq_Ws_get_Omega_Lambda0(W_A, kap_A, H_A, W_C, kap_C, H_C, K_ratio = 1,
  I_P, pars)
}
\arguments{
\item{W_A}{equilibrium mean worm burden in adult population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{kap_A}{equilibrium clumping parameter in adult population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{H_A}{adult human population size}

\item{W_C}{equilibrium mean worm burden in child population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{kap_C}{equilibrium clumping parameter in child population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{H_C}{child human population size}

\item{K_ratio}{ratio of snail carrying capacity to human population size. defaults to 1, but should be increased for environments/communities that have particularly intense snail habitat or tranmission}

\item{I_P}{infected snail prevalence, observed or input}

\item{pars}{additional model parameters}
}
\value{
Estimates of the relative child to adult exposure/contamination ratio
}
\description{
Function which takes input infection and demographic information and estimates relative infection between adults and children
assuming equilibrium differences in worm burden are a result of this parameter. Then estimates peak miracidial invasion rate
from a resulting estimate of the equilibrium miracidial density as a function of worm burden, contamination coefficients and other parameters
}
