% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{eqbm_snails_fit_mod}
\alias{eqbm_snails_fit_mod}
\title{Solve for equilibrium susceptible snail and total snail population}
\usage{
eqbm_snails_fit_mod(x, pars, W_TC, W_UC, W_TA, W_UA)
}
\arguments{
\item{x}{input estimates of eqbm population sizes, N and S (IN THAT ORDER) for initial estimates}

\item{pars}{parameter set}

\item{W_TC}{mean worm burden in treated SAC group}

\item{W_UC}{mean worm burden in untreated SAC group}

\item{W_TA}{mean worm burden in treated adult group}

\item{W_UA}{mean worm burden in untreated adult group}
}
\value{
vector of estimates of total equlibirum snail population size and proportion susceptible, exposed, and infected
}
\description{
Takes as input the worm burden in each age and treatment group and
vector of model parameters and uses `rootSolve::multiroot` to determine the
total, N, and susceptible, S, snail population sizes. Returns these values along
with resulting estimates of exposed, E, and infected, I, snail populations.
These values returned as ratios, i.e. S/N; E/N; I/N
}
