% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{exp_seq}
\alias{exp_seq}
\title{Function to generate log transformed sequence evenly distributed across broad orders of magnitude}
\usage{
exp_seq(min, max, seq.length)
}
\arguments{
\item{min}{minimum value in the sequence}

\item{max}{maximum value in the sequence}

\item{seq.length}{length of the sequence}
}
\value{
numeric vector spanning min and max with n = seq.length entries
}
\description{
Functions just like `seq` but evenly distributes values across the full range
rather than for instance `seq(0.000001, 10000, length.out = 100)` returning values that are all >100
}
