% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{f_Wgk}
\alias{f_Wgk}
\title{Schistosomiasis fecundity reduction due to crowding at high densities}
\usage{
f_Wgk(W, gamma, k)
}
\arguments{
\item{W}{Mean worm burden in human population}

\item{gamma}{shape parameter regulating response of egg ouptut to worm burden}

\item{k}{clumping parameter of the negative binomial distribution}
}
\value{
Proportional reduction in female worm egg output due to crowding
}
\description{
Reductions in egg output due to crowding of adult female worms at high worm burdens
}
