% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{fit_pars_from_eq_vals}
\alias{fit_pars_from_eq_vals}
\title{Uses `W_Ip_eq_solns` within `multiroot` to solve for unkown transmission parameters}
\usage{
fit_pars_from_eq_vals(beta_guess, alpha_guess, Neq_guess, W, Ip, pars)
}
\arguments{
\item{beta_guess}{initial estimate for snail FOI parameter, beta}

\item{alpha_guess}{initial estimate for human FOI parameter, alpha}

\item{Neq_guess}{initial estimate for equlibirum snail population size}

\item{W}{input equlibirum mean worm burden}

\item{Ip}{input equlibirum infected snail prevalence}

\item{pars}{additional model parameters}
}
\value{
vector of alpha, beta, and N_eq estimates
}
\description{
Uses euqlibirum equations for snail population, mean worm burden, and snail FOI
and input equilibirum mean worm burden and infected snail prevalence to solve for unkown transmission
parameters and infected snail prevalence
}
