% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{gam_Wxi}
\alias{gam_Wxi}
\title{Schistosomiasis acquired immunity}
\usage{
gam_Wxi(W, xi)
}
\arguments{
\item{W}{mean worm burden in human population}

\item{xi}{density dependence shape parameter regulating response of immunity effect to mean worm burden}
}
\value{
reduction in snail-to-man infection as a result of acquired immunity in human population
}
\description{
Reductions in transmission at high worm burdens due to acquired immunity in human population
}
