% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{garch_discrete_mod_dd}
\alias{garch_discrete_mod_dd}
\title{Garchitorena discrete time simulator with density dependence}
\usage{
garch_discrete_mod_dd(I, p)
}
\arguments{
\item{I}{value of the state variable I at time t}

\item{p}{parameter set}
}
\value{
value of I at t+1
}
\description{
Same model as in `garch_mod_dd`, but discrete time version for
ease of implementation with stochastic dynamic programing framework
}
