% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{garch_mod}
\alias{garch_mod}
\title{Garchitorena et al model for environmentally transmitted infectious diseases}
\usage{
garch_mod(t, n, p)
}
\arguments{
\item{t}{Vector of timepoints to return state variable estiamtes}

\item{n}{Vector of state variable initial conditions}

\item{p}{Named vector of model parameters}
}
\value{
A matrix of the state variables at all requested time points
}
\description{
Deterministic version of the simple, two-differential-equation model
presented in Garchitorena et al 2018
}
