% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{garch_mod_dd}
\alias{garch_mod_dd}
\title{Garchitorena et al model for environmentally transmitted infectious diseases
augmented with density dependent function}
\usage{
garch_mod_dd(t, n, p)
}
\arguments{
\item{t}{Vector of timepoints to return state variable estiamtes}

\item{n}{Vector of state variable initial conditions}

\item{p}{Named vector of model parameters}
}
\value{
A matrix of the state variables at all requested time points
}
\description{
Deterministic version of the simple, two-differential-equation model
presented in Garchitorena et al 2018 with a function dependent on
prevalence (I) that represents a density dependence in transmission
and gives rise to a transmission breakpoint and endemic equilibrium
}
