% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{garch_r0}
\alias{garch_r0}
\title{Garchitorena et al $R_0$ estimation}
\usage{
garch_r0(p)
}
\arguments{
\item{p}{parameter set}
}
\value{
estimate of $R_0$
}
\description{
For the simple, two-equation infectious disease model in
`garch_mod`, estimate basic reproduction number, R_0 for the given parameter set
}
