% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{gen_force_fx}
\alias{gen_force_fx}
\title{Generate seasonal model forcing function}
\usage{
gen_force_fx(burn_in, t_int, t_max, freq, par_set, par, par_change)
}
\arguments{
\item{burn_in}{time period in the beginning to leave parameter unaffected if desired}

\item{t_int}{time period during which intervention occurs}

\item{t_max}{total time frame of simulation as numeric}

\item{freq}{frequency of the seasonality}

\item{par_set}{named vector of the parameter set that contains the parameter affected}

\item{par}{character of the parameter in the parameter set that is affected}

\item{par_change}{relative amplitude of the seasonal change in the parameter value}
}
\value{
Forcing function
}
\description{
Function to generate a seasonal forcing function based on the simulation time frame
the parameter affected, and the phase of the seasonality.
time_frame/freq should be a whole number indicating the number of seasonal cycles
}
