% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{get_breakpoint_W_N}
\alias{get_breakpoint_W_N}
\title{Worm burden breakpoint estimation}
\usage{
get_breakpoint_W_N(Wbp_guess, Nbp_guess, pars)
}
\arguments{
\item{Wbp_guess}{initial value of Wbp to use}

\item{Nbp_guess}{initial value of N(Wbp) to use}

\item{pars}{additional parmaters}
}
\value{
vector with Wbp and N(Wbp)
}
\description{
Uses `multiroot` and `wbp_fx` functions to estimate Wb and N(Wbp)
}
