% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_stratified_reproduction_numbers.R
\name{get_Ip_W_bp}
\alias{get_Ip_W_bp}
\title{Get breakpoint snail infection prevalence and mean worm burden}
\usage{
get_Ip_W_bp(N_guess, W_bp_guess, parms)
}
\arguments{
\item{N_guess}{initial guess for breakpoint snail population size}

\item{W_bp_guess}{initial guess for breakpoint mean worm burden}

\item{parms}{parameter values}
}
\description{
Uses `multiroot` to get estimates of snail population size and
breakpoint mean worm burden from equations for snail population dynamics and worm burden estimation
then uses these estimates to determine breakpoint snail prevalence
}
