% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{get_R0}
\alias{get_R0}
\title{Estimate R_0 from model parameters}
\usage{
get_R0(pars)
}
\arguments{
\item{pars}{parameters}
}
\value{
estimate of R0
}
\description{
Estimate R_0 from model parameters
}
