% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{getReff_noPDD}
\alias{getReff_noPDD}
\title{Get estimate of effective reproduction number, $R_eff$ absen influence of positive density dependence}
\usage{
getReff_noPDD(parameters, W, kap)
}
\arguments{
\item{parameters}{Model parameters}

\item{W}{mean worm burden in human population}

\item{kap}{clumping parameter of the worm burden among human population}
}
\value{
estimate of the effective reproduction number, $R_eff$
}
\description{
Estimates the effective reproduction number, $R_eff$ as a function of
model parameters, population mean worm burden, and clumping parameter of negative binomial worm burden distribution
in the absence of density dependence expressed as the mating probability
}
