% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{I_get_E}
\alias{I_get_E}
\title{Estimate exposed snail fraction from infected snail prevalence}
\usage{
I_get_E(I, pars)
}
\arguments{
\item{I}{infected snail prevalence}

\item{pars}{parameter set}
}
\value{
Estimate of porportion of snail population in E compartment
}
\description{
Function to estimate proportion of snail population in exposed, E, compartment from model parameters
and infected snail, I, prevalence
}
