% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{I_get_Lambda}
\alias{I_get_Lambda}
\title{Function to estimate Lambda (snail-to-man FOI) as function of snail infection prevalence and snail parameters}
\usage{
I_get_Lambda(I_P, mu_N, mu_I, sigma)
}
\arguments{
\item{I_P}{infected snail prevalence}

\item{mu_N}{snail daily mortality rate}

\item{mu_I}{infected snail daily mortality rate}

\item{sigma}{inverse of snail pre-patency period}
}
\value{
Estimate of man-to-snail FOI
}
\description{
Snail FOI can be estimated as a function of observed infected snail prevalence and additional snail parameters
This is a function to estimate this parameter from input infected snail prevalence and model parameters.
}
