% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{I_get_lambda0}
\alias{I_get_lambda0}
\title{Estimate peak snail infection from input snail infection prevalence}
\usage{
I_get_lambda0(I, pars)
}
\arguments{
\item{I}{infected snail prevalence}

\item{pars}{parameter set}
}
\value{
Estimate of porportion of snail population in E compartment
}
\description{
Peak infection rate can be estimated as a function of observed infected snail prevalence as in
www.pnas.org/cgi/doi/10.1073/pnas.1708729114. This is a function to estimate this parameter
from input infected snail prevalence and model parameters. Depends on function `I_get_E` as well
}
