% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{I_P_get_W_eq}
\alias{I_P_get_W_eq}
\title{Estimate endemic equilibrium given infected snail prevalence}
\usage{
I_P_get_W_eq(I_P, pars)
}
\arguments{
\item{I_P}{infected snail prevalence}

\item{pars}{named vector or list of other parameters}
}
\value{
estimate of the breakpoint
}
\description{
Using Reff expression and fact that Reff=1 at the endemic equilibrium, we can express the endemic equilibirum in terms of
the man-to-snail FOI, Lambda. We can also express Lambda in terms of the infected snail prevalence, I_P,
therefore we can estimate the endemic equilibrium as a function of model parameters and input snail prevalence
}
