% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{infection_inputs_get_pars}
\alias{infection_inputs_get_pars}
\title{Function to input observed equilibrium infection values and return fully "fit" parameter set based on observed infection values}
\usage{
infection_inputs_get_pars(W_A, kap_A, H_A, cvrg_A, W_C, kap_C, H_C, cvrg_C,
  K_ratio = 1, I_P, pars)
}
\arguments{
\item{W_A}{equilibrium mean worm burden in adult population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{kap_A}{equilibrium clumping parameter in adult population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{H_A}{adult human population size}

\item{cvrg_A}{MDA coverage of the adult population}

\item{W_C}{equilibrium mean worm burden in child population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{kap_C}{equilibrium clumping parameter in child population, can be estimated from egg burden and prevalence using `convert_burden_egg_to_worm`}

\item{H_C}{child human population size}

\item{cvrg_C}{MDA coverage of the child population}

\item{K_ratio}{ratio of snail carrying capacity to human population size. defaults to 1, but should be increased for environments/communities that have particularly intense snail habitat or tranmission}

\item{I_P}{infected snail prevalence, observed or input}

\item{pars}{additional model parameters}
}
\value{
parameter set with fit parameters based on equilibirum values
}
\description{
Uses `eq_W_get_alpha`, `eq_Ws_get_Omega_Lambda0`, `convert_burden_egg_to_worm`, `Lambda_get_N_eq`, and `I_get_Lambda` to convert input
infection values (mean egg burden and infection prevalence in adult and child populations, infected snail prevalence) into estimates of model parameters
}
