% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{k_from_log_W}
\alias{k_from_log_W}
\title{Estimate clumping parameter as a function of log worm burden}
\usage{
k_from_log_W(W)
}
\arguments{
\item{W}{Mean worm burden or infection intensity}
}
\value{
Estimate of the clumping parameter of the negative binomial distribution
}
\description{
Function fit to our data from 16 communities in Senegal with annual MDA
and egg burden assessed annually for three years
}
