% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{Lambda_get_N_eq}
\alias{Lambda_get_N_eq}
\title{Function to estimate equilibirum snail population size, N_eq, as function of Lambda and snail parameters}
\usage{
Lambda_get_N_eq(Lambda, K, mu_N, r, sigma)
}
\arguments{
\item{K}{snail population environmental carrying capacity}

\item{mu_N}{snail daily mortality rate}

\item{r}{snail intrinsic reproduction rate}

\item{sigma}{inverse of snail pre-patency period}

\item{I_P}{infected snail prevalence}
}
\value{
Estimate of man-to-snail FOI
}
\description{
Total equilibrium snail population can be estimated as a function of Lambda, the man-to-snail FOI and other parameters.
This function takes as input Lambda (which can be estimated as a function of infected snail prevalence using function `I_get_Lambda`)
and additional snail parameters to estimate N_eq
}
