% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{Lambda_Wij_linear}
\alias{Lambda_Wij_linear}
\title{Estimate linear form of man-to-snail FOI}
\usage{
Lambda_Wij_linear(pars, W_TC, W_UC, W_TA, W_UA)
}
\arguments{
\item{pars}{parameter set}

\item{W_TC}{mean worm burden in treated SAC group}

\item{W_UC}{mean worm burden in untreated SAC group}

\item{W_TA}{mean worm burden in treated adult group}

\item{W_UA}{mean worm burden in untreated adult group}
}
\value{
Estimate of the man-to-snail FOI from the linear function
}
\description{
Estimates the linear version of man-to-snail FOI as function of
model parameters and mean worm burden in each age and treatment population
}
