% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{latexImg}
\alias{latexImg}
\title{Function to insert latex equations into .md dox}
\usage{
latexImg(latex)
}
\arguments{
\item{latex}{latex code to generate equation}
}
\description{
Function copied directly from here:
https://github.com/STAT545-UBC/Discussion/issues/102
equations don't render properly
}
