% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{nil_1}
\alias{nil_1}
\title{Return 1}
\usage{
nil_1(...)
}
\value{
1
}
\description{
Function to take any input and return 1, used as a placeholder in some functions where input density dependent functions are required
}
