% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{phi_Wk}
\alias{phi_Wk}
\title{Schistosomiasis mating probability}
\usage{
phi_Wk(W, k)
}
\arguments{
\item{W}{mean population worm burden}

\item{k}{clumping parameter of the negative binomial distribution}
}
\value{
probability a female worm is succefully mated (ranges from 0-1)
}
\description{
Estimate the mating probability from the population mean worm burden and the
clumping parameter of the negative binomial distribution
}
