% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{prev_get_W}
\alias{prev_get_W}
\title{Estimate mean worm burden and clumping parameter as function of prevalence}
\usage{
prev_get_W(Prev)
}
\arguments{
\item{Prev}{Prevalence in th epopulation}
}
\value{
Estimate of the mean worm burden
}
\description{
This is based on the assumption that the clumping parameter is determined by mean worm burden
which reduces the number of unkowns to 2 therefore if prevalence is known, mean worm burden
and kappa as determined by the mean worm burden are known
}
