% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{prev_intensity_to_clump}
\alias{prev_intensity_to_clump}
\title{Estimate clumping parameter as function of prevalence and intensity}
\usage{
prev_intensity_to_clump(W, Prev)
}
\arguments{
\item{W}{Mean worm burden or infection intensity}

\item{Prev}{Prevalence in th epopulation}
}
\value{
Estimate of the clumping parameter
}
\description{
Prevalence, mean intensity and the clumping parameter are all related,
therefore estimation of 1 can be achieved if the other two are known
}
