% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{prev_kap_get_W}
\alias{prev_kap_get_W}
\title{Estimate mean worm burden from egg prevalence and clumping parameter}
\usage{
prev_kap_get_W(prev, kap)
}
\value{
Estimate of the mean worm burden
}
\description{
@param prev prevalence of signs of infection (e.g. eggs in urine)
 @param kap dispersion parameter of infection intensity
}
