% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{prev_W_get_k}
\alias{prev_W_get_k}
\title{Estimate clumping parameter, kappa, as function of mean infection intensity and prevalence using uniroot}
\usage{
prev_W_get_k(W, prev)
}
\arguments{
\item{W}{Mean worm burden or infection intensity}

\item{prev}{prevalence of infection}
}
\value{
Estimate of the clumping parameter
}
\description{
Prevalence, mean intensity and the clumping parameter are all related,
therefore estimation of 1 can be achieved if the other two are known
}
