% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{R0_egg_kap_zeta_m}
\alias{R0_egg_kap_zeta_m}
\title{Estimate R_0 from input egg burden and density dependence parameters}
\usage{
R0_egg_kap_zeta_m(eggs, kap, zeta, m)
}
\arguments{
\item{eggs}{mean population egg burden}

\item{kap}{dispersion parameter}

\item{zeta}{negative density dependence parameter}

\item{m}{peak egg output}
}
\value{
estimate of basic reproduction number, R0
}
\description{
Estimate R_0 from input egg burden and density dependence parameters
}
