% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{R_eff_W_I}
\alias{R_eff_W_I}
\title{Estimate effective reproduction number as product of W, I and parameters}
\usage{
R_eff_W_I(W, I, pars, PDD = phi_Wk, DDF = rho_Wk, DDI = nil_1)
}
\arguments{
\item{W}{mean worm burden}

\item{I}{infected snail population size}

\item{pars}{additional parmaters}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}
}
\value{
vector with Reff and R0 estimates
}
\description{
Takes mean worm burden from unstratified population, infected snail population size
and additional parameters and returns R_eff estimate
}
