% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{R_eff_W_I_P}
\alias{R_eff_W_I_P}
\title{Estimate effective reproduction number as product of W, I_P and parameters}
\usage{
R_eff_W_I_P(W, I_P, pars)
}
\arguments{
\item{W}{mean worm burden}

\item{I_P}{infected snail prevalence}

\item{pars}{additional parmaters}
}
\value{
vector with Reff and R0 estimates
}
\description{
Takes mean worm burden from unstratified population, infected snail prevalence
and additional parameters and returns both R_eff and R_0 estimates
}
