% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{R_Wv}
\alias{R_Wv}
\title{Schistosomiasis acquired immunity}
\usage{
R_Wv(W, v)
}
\arguments{
\item{W}{mean worm burden in human population}

\item{v}{density dependence shape parameter regulating response of immunity effect to mean worm burden}
}
\value{
reduction in snail-to-man infection as a result of acquired immunity in human population
}
\description{
Reductions in transmission at high worm burdens due to acquired immunity in human population
}
