% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{Reff_W}
\alias{Reff_W}
\title{Get Reff as function of W, model parameters, and DDs}
\usage{
Reff_W(W, pars, PDD = phi_Wk, DDF = rho_Wk, DDI = gam_Wxi)
}
\arguments{
\item{W}{mean worm burden}

\item{pars}{other model parameters}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}
}
\value{
estimate of Reff
}
\description{
Get Reff as function of W, model parameters, and DDs
}
