% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{Reff_Wij}
\alias{Reff_Wij}
\title{Estimate effective reproduction number}
\usage{
Reff_Wij(pars, W_TC, W_UC, W_TA, W_UA)
}
\arguments{
\item{pars}{parameter set}

\item{W_TC}{mean worm burden in treated SAC group}

\item{W_UC}{mean worm burden in untreated SAC group}

\item{W_TA}{mean worm burden in treated adult group}

\item{W_UA}{mean worm burden in untreated adult group}
}
\description{
Estimates the effective reproduction number, $R_eff$ as a function of
model parameters and mean worm burden in each age and treatment population
}
