% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_stratified_reproduction_numbers.R
\name{Reff_Wij_I_P}
\alias{Reff_Wij_I_P}
\title{Estimate effective reproduction number as function of mean worm burden in different strata, infected snail prevalence and other parameters}
\usage{
Reff_Wij_I_P(pars, W_TC, W_UC, W_TA, W_UA, I_P, PDD, DDF, DDI)
}
\arguments{
\item{pars}{parameter set}

\item{W_TC}{mean worm burden in treated SAC group}

\item{W_UC}{mean worm burden in untreated SAC group}

\item{W_TA}{mean worm burden in treated adult group}

\item{W_UA}{mean worm burden in untreated adult group}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}
}
\value{
estimate of the effective reproduction number and the population average mean worm burden
}
\description{
Estimates the effective reproduction number, $R_eff$ as a function of
model parameters, mean worm burden in each age and treatment population, infected snail prevalence, and
different options for density dependence parameters
}
