% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{schisto_age_strat_mod}
\alias{schisto_age_strat_mod}
\title{Age-stratified schistosomiasis model}
\usage{
schisto_age_strat_mod(t, n, parameters)
}
\arguments{
\item{t}{Vector of timepoints to return state variable estiamtes}

\item{n}{Vector of state variable initial conditions}

\item{parameters}{Named vector of model parameters}
}
\value{
A matrix of the state variables at all requested time points
}
\description{
An age-stratified dynamic schistosomiasis model with treated and untreated
school-aged children (SAC) and adult populations,
negative (crowding induced reductions in fecundity)
and positive (mating limitation) density dependencies and functionality
to simulate mass drug administration, snail control, and other interventions.
Also incorporates non-linear snail-to-man transmission dynamics as derived in Gurarie et al 2018
Note this is a function that is wrapped into `sim_schisto_age_strat_mod`` which
should be used to simulate the model, this function is fed into the ode solver
from deSolve
}
