% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_dynamic_models.R
\name{schisto_mod}
\alias{schisto_mod}
\title{Barebones schistosomiasis model}
\usage{
schisto_mod(t, n, parameters)
}
\arguments{
\item{t}{Vector of timepoints to return state variable estiamtes}

\item{n}{Vector of state variable initial conditions}

\item{parameters}{Named vector of model parameters}
}
\value{
A matrix of the state variables at all requested time points
}
\description{
A dynamic schistosomiasis model with SEI snail infection dynamics, a single mean worm burden population,
negative (crowding induced reductions in fecundity)
and positive (mating limitation) density dependencies and functionality
to simulate mass drug administration, snail control, and other interventions.
Note this is a function that is wrapped into `sim_schisto_base_mod`` which
should be used to simulate the model, this function is fed into the ode solver
from deSolve
}
