% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{sim_garch_mod}
\alias{sim_garch_mod}
\title{Simulate the garchitorena model}
\usage{
sim_garch_mod(nstart, time, model = garch_mod, parameters,
  events_df = NA)
}
\arguments{
\item{nstart}{Named vector of starting values for state variables}

\item{time}{Numeric vector of times at which state variables should be estimated}

\item{model}{Name of the ode function to use, defaults to `garch_mod`}

\item{parameters}{Named vector or list of parameter values}

\item{events_df}{Data frame of events such as MDA with columns "var", "time", "value", and "method"}
}
\value{
dataframe of state variable values at requested times
}
\description{
Simulate garchitorena deterministic model through time using `ode` function
given the model deSolve function, parameter set, time frame, starting conditions, and potential events (e.g. MDA)
}
