% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{sim_int_choice}
\alias{sim_int_choice}
\title{Garchitorena discrete time model simulation with intervention variable}
\usage{
sim_int_choice(A_t, I_0, int_par, time = 365, parameters)
}
\arguments{
\item{A_t}{decision variable: proportion of capital to allocate towards MDA}

\item{I_0}{Initial value of prevalence}

\item{int_par}{parameter for remaining capital intervention to be allocated towards}

\item{time}{Total time (days) to simulate}

\item{parameters}{Named vector or list of parameter values}
}
\value{
value of prevalence state variable at the end of the time period of intervention
}
\description{
Simulates discrete time version of the garchitorena model
builds on `garch_discrete_mod` and incorporates decision variable A_t
that regulates where capital is committed to intervention
}
