% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{sim_opt_choice}
\alias{sim_opt_choice}
\title{Implement an optimal policy identified with MDPtoolbox through time}
\usage{
sim_opt_choice(A_vec, states, opt_list, p_start, t_per_step, int_par,
  parameters)
}
\arguments{
\item{A_vec}{vector of possible decisions}

\item{states}{vector of possible states of the system}

\item{opt_list}{List from `MDPtoolbox::mdp_finite_horizon`}

\item{p_start}{index of which state to start at}

\item{t_per_step}{Total time (days) to simulate}

\item{int_par}{parameter environmental intervention acts on}

\item{parameters}{Named vector or list of parameter values}
}
\value{
Data frame of prevalence over course of intervention with optimal intervention implemented at each time step
}
\description{
Simulates optimal treatment allocation based on control variable A_t
in the discrete time version of the garchitorena model
builds on `garch_discrete_mod` and incorporates decision variable A_t
that regulates where capital is committed to intervention
}
