% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_dynamic_models.R
\name{sim_schisto_base_mod}
\alias{sim_schisto_base_mod}
\title{Simulate the base schistosomiasis model}
\usage{
sim_schisto_base_mod(nstart, time, model = schisto_base_mod, parameters)
}
\arguments{
\item{nstart}{Named vector of starting values for state variables}

\item{time}{Numeric vector of times at which state variables should be estimated}

\item{model}{Name of the ode function to use, defaults to `schisto_base_mod`}

\item{parameters}{Named vector or list of parameter values}
}
\value{
dataframe of state variable values at requested times
}
\description{
Uses the `schisto_base_mod` function to simulate the model through time
given parameter set, time frame, starting conditions, and potential events (e.g. MDA)
}
