% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_dynamic_models.R
\name{sim_schisto_mod}
\alias{sim_schisto_mod}
\title{Simulate a schistosomiasis model}
\usage{
sim_schisto_mod(nstart, time, model = schisto_base_mod, parameters,
  events_df = NA)
}
\arguments{
\item{nstart}{Named vector of starting values for state variables}

\item{time}{Numeric vector of times at which state variables should be estimated}

\item{model}{Name of the ode function to use, defaults to `schisto_base_mod`}

\item{parameters}{Named vector or list of parameter values}

\item{events_df}{Data frame of events such as MDA with columns "var", "time", "value", and "method"}
}
\value{
dataframe of state variable values at requested times
}
\description{
Simulate a schisto deterministic model through time using `ode` function
given the schisto model deSolve function, parameter set, time frame, starting conditions, and potential events (e.g. MDA)
}
