% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_dynamic_models.R
\name{sim_schisto_stoch_mod}
\alias{sim_schisto_stoch_mod}
\title{Simulate the stochastic schistosomiasis model}
\usage{
sim_schisto_stoch_mod(nstart, transitions = list(c(S = 1), c(S = -1), c(S
  = -1, E = 1), c(E = -1), c(E = -1, I = 1), c(I = -1), c(Wt = 1, Wu = 1),
  c(Wt = -1), c(Wu = -1)), sfx = schisto_stoch_mod, params, tf,
  events_df = NA)
}
\arguments{
\item{nstart}{starting values of state variables, must be whole numbers}

\item{transitions}{list of all possible transitions between state variables, defaults to transitions possible in `schisto_stoch_mod`}

\item{sfx}{function to feed into adaptive tau, uses `schisto_stoch_mod` as default}

\item{params}{named vector of parameter values used in model}

\item{tf}{numeric of total time to run simulation}

\item{events_df}{Data frame of events such as MDA with columns "var", "time", "value", and "method"}
}
\value{
matrix of state variables and values at time points up until tf
}
\description{
Uses the `schisto_stoch_mod` function to simulate the stochastic model through time
}
