% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{sim_w_inputs}
\alias{sim_w_inputs}
\title{Simulate transmission and intervention through time with variable inputs}
\usage{
sim_w_inputs(R0, script_P, T_frame, freq, M, A, pars, mod)
}
\arguments{
\item{R0}{transmission intensity}

\item{script_P}{relative cost of environmental to human intervention}

\item{T_frame}{total time frame to simulate}

\item{freq}{frequency of intervention}

\item{M}{total capital available per intervention}

\item{A}{numeric between 0 and 1 of proportion of capital to go towards human intervention (remainder goes towards environmental intervention)}

\item{pars}{parameter set with remainin transmission parameters}

\item{mod}{model formulation (e.g. with density dependence `garch_mod_pdd` or not `garch_mod`)}
}
\value{
estimate of utility associated with inputs. Measured as the sum over time of prevalence^1.5 (penalizes higher prevalence values)
}
\description{
For the simple, two-equation infectious disease model in
`garch_mod`, simulate transmission and intervention with choices
for transmission intensity (R0), relative cost of environmental to human
intervention (script_P), timeframe, capital available, allocation of capital,
parameters, and model formulation
}
