% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{W_bp}
\alias{W_bp}
\title{Estimate breakpoint as function of I_P and parameters}
\usage{
W_bp(pars, PDD, DDF, DDI)
}
\arguments{
\item{pars}{additional parmaters}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}

\item{I_P}{infected snail prevalence}
}
\value{
vector with Reff and R0 estimates
}
\description{
Estimate breakpoint as function of I_P and parameters
}
