% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_stratified_reproduction_numbers.R
\name{W_bp_N_solver}
\alias{W_bp_N_solver}
\title{Model to feed to `rootSolve::multiroot()` to estimate breakpoint from input parameters}
\usage{
W_bp_N_solver(x, parms)
}
\arguments{
\item{x}{input estimates of values N_eq and W_bp (IN THAT ORDER)}

\item{parms}{parameter set of other key parameters}
}
\value{
Vector of two solutions
}
\description{
Uses the equilibrium snail population and reff equations to estimate equilibirum snail population size
and population breakpoint. This function takes estimates of these two values and returns values of these equations.
Best use is within `multiroot` to find roots which represent the snail population sizr and mean worm burden at the breakpoint
}
