% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{W_bp_R_0}
\alias{W_bp_R_0}
\title{Estimate W_bp from R_0}
\usage{
W_bp_R_0(R_0, PDD, DDF, DDI, gamma, xi)
}
\arguments{
\item{R_0}{basic reproduction number}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}

\item{gamma}{DDF parameter}

\item{DDI}{parameter}
}
\value{
estimate of the breakpoint mean worm burden
}
\description{
Estimate W_bp from R_0
}
