% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_age_structured_models.R
\name{W_ij_get_N_S_E_I}
\alias{W_ij_get_N_S_E_I}
\title{Estimate Snail infection class distribution given mean worm burden in different infection/treatment classes}
\usage{
W_ij_get_N_S_E_I(W_TC, W_UC, W_TA, W_UA, pars)
}
\arguments{
\item{W_TC}{mean worm burden in treated SAC group}

\item{W_UC}{mean worm burden in untreated SAC group}

\item{W_TA}{mean worm burden in treated adult group}

\item{W_UA}{mean worm burden in untreated adult group}

\item{pars}{named vector or list of other parameters}
}
\value{
estimate of N, S, E, and I
}
\description{
Takes mean worm burden estimates in different worm burden classes and estimates miraicidial concentration
Then uses snail population dynamics equation as a function of LINEAR snail FOI to estimate snail population size
Then uses parameters to estimate infection class distribution from total snail population size
}
