% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{W_Ip_eq_solns}
\alias{W_Ip_eq_solns}
\title{Function used in `multiroot`` to fit transmission parameters}
\usage{
W_Ip_eq_solns(x, parms, W, Ip)
}
\arguments{
\item{x}{starting values of unkowns, a vector of beta, alpha, and N_eq}

\item{parms}{additional model parameters}

\item{W}{equilibirum mean worm burden}

\item{Ip}{equlibirum infected snail prevalence}
}
\value{
vector of solutions to three equations from inputs
}
\description{
Fits transmission parameters, alpha and beta, based on equilibrium solutions to snail and worm burden equations
Given unkown beta, alpha, and equilibirum snail population, N_eq
}
