% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{w_start_get_r0}
\alias{w_start_get_r0}
\title{Estimate R0 from starting worm burden and density dependence parameters}
\usage{
w_start_get_r0(w_start, kap, gamma)
}
\arguments{
\item{w_start}{endemic equilibrium mean worm burden}

\item{kap}{endemic equilibrium dispersion parameter}

\item{gamma}{negative densiy dependent fecundity parameter}
}
\value{
estimate of basic reproduction number, R0
}
\description{
Estimate R0 from starting worm burden and density dependence parameters
}
