% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{wbp_fx}
\alias{wbp_fx}
\title{Worm burden breakpoint equations for use in multiroot}
\usage{
wbp_fx(x, parms)
}
\arguments{
\item{x}{vector of the two state variables, Wbp and N(Wbp)}

\item{parms}{additional parmaters}
}
\value{
vector with equation estimates from Wbp and N(Wbp)
}
\description{
Equations for worm burden breakpoint and snail population at the breakpoint used in `multiroot` to identify roots
which represent the breakpoint and snail population at the breakpoint
}
