% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_reproduction_numbers.R
\name{cvrg_from_pars}
\alias{cvrg_from_pars}
\title{Estimate MDA coverage necessary to reach breakpoint as a function of parameters
used to esimate breakpoint, density dependencies,}
\usage{
cvrg_from_pars(use_pars, PDD = phi_Wk, DDF = rho_Wk, DDI = nil_1)
}
\arguments{
\item{use_pars}{list of parameters necessary to estimate breakpoint and coverage including
mean pre-treatment worm burden in treated population, W_t, mean pre-treatment worm burden in the untreated population, W_u, drug efficacy, epsilon, snail population environmental carrying capacity, K, and sanitation/exposure parameter, omega, and other model parameters, pars}

\item{PDD}{positive density dependence function}

\item{DDF}{density dependent fecundity function}

\item{DDI}{density dependent acquired immunity function}
}
\value{
estimate of coverage required to reach breakpoint
}
\description{
Estimate MDA coverage necessary to reach breakpoint as a function of parameters
used to esimate breakpoint, density dependencies,
}
