% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schisto_helper_functions.R
\name{est_prev_W_k}
\alias{est_prev_W_k}
\title{Estimate prevalence as function of clumping parameter and mean infection intensity}
\usage{
est_prev_W_k(W, k)
}
\arguments{
\item{W}{Mean worm burden or infection intensity}

\item{k}{clumping parameter of the negative binomial distribution}
}
\value{
Estimate of the prevalence
}
\description{
Prevalence, mean intensity and the clumping parameter are all related,
therefore estimation of 1 can be achieved if the other two are known
}
