% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchitorena_models.R
\name{sim_discrete_mod}
\alias{sim_discrete_mod}
\title{Garchitorena discrete time model simulation}
\usage{
sim_discrete_mod(I_0, time, model = garch_discrete_mod, parameters,
  events_df = NA)
}
\arguments{
\item{I_0}{Initial value of prevalence}

\item{time}{Total time (days) to simulate}

\item{model}{Name of the discrete sim function to use, defaults to `garch_discrete_mod`}

\item{parameters}{Named vector or list of parameter values}

\item{events_df}{Data frame of events such as MDA with columns "var", "time", "value" where var is the variable (state variable or parameter) affected, time is the time to implement the event, and value is the value of the change in the value of the variable or parameter as a proportion (e.g. between 0 and 1)}
}
\value{
dataframe of state variable values at requested times
}
\description{
Simulates discrete time version of the garchitorena model
builds on `garch_discrete_mod`
}
